+++ Tree.py
@@ -1,0 +1,12 @@
+class Tree(object):
+
+    def __init__(self, c1, c2, c3, c4, c5, c6, kc, desc, number):
+        self.c1 = c1
+        self.c2 = c2
+        self.c3 = c3
+        self.c4 = c4
+        self.c5 = c5
+        self.c6 = c6
+        self.kc = kc
+        self.desc = desc
+        self.number = number
