+++ spriteSheet.py
@@ -1,0 +1,32 @@
+import pygame
+from collections import deque
+from random import randint
+
+class spriteSheet(object): #create spriteSheet class to handle sprite sheets
+    def __init__(self, filename, cols, rows): #constructor with arguments filename, rows, and columns
+        self.sheet = pygame.image.load(filename).convert_alpha() #load the spritesheet
+        
+        self.cols = cols 
+        self.rows = rows
+        self.totalCellCount = cols * rows #get total number of cells based on rows/columns
+
+        self.rect = self.sheet.get_rect() #get the area of the sprite sheet
+        w = self.cellWidth = self.rect.width / cols #each cell's width is the whole sheet's width divided by the number of columns
+        h = self.cellHeight = self.rect.height / rows #each cell's height is the whole sheet's height divided by the number of rows
+
+
+    def getSpriteList(self):
+        x = 0
+        y = 0
+        spriteList = [] #initialize list to hold sprite coordinates in
+        for row in range(0, self.rows): #for each row in the sprite sheet:
+            for column in range(0, self.cols): #for each column in the sprite sheet:
+                coords = (x, y) #create tuple of the x and y from above
+                spriteList.append(coords) #append to list
+                x -= self.cellWidth #increment x by the width of each cell 
+            y -= self.cellHeight #then when going to the next row, increment the y by the height of a cell
+        return spriteList #return the list
+
+        
+    def draw(self, surface, x, y):
+        surface.blit(self.sheet, (x, y, self.cellWidth, self.cellHeight))
